Test&Feature : FeePolicy ( 가격 정책 관련 객체 생성 ) 및 Sections 에서 담당하던 관련 테스트 이관
